Renaming news-fetcher to news-fetcher-service
