+++ back/news-fetcher-service/.gitattributes
