+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/NewsFetcherApplication.java
@@ -1,0 +1,15 @@
+package com.ds.news_fetcher_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class NewsFetcherApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NewsFetcherApplication.class, args);
+	}
+
+}
