+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/config/RedisConfig.java
@@ -1,0 +1,22 @@
+package com.ds.news_fetcher_service.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
+import org.springframework.data.redis.core.ReactiveRedisTemplate;
+import org.springframework.data.redis.serializer.RedisSerializationContext;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+
+@Configuration
+public class RedisConfig {
+    @Bean
+    public ReactiveRedisTemplate<String, String> reactiveRedisTemplate(ReactiveRedisConnectionFactory factory){
+        RedisSerializationContext<String, String> context  = RedisSerializationContext
+                .<String, String>newSerializationContext(new StringRedisSerializer())
+                .value(new StringRedisSerializer())
+                .build();
+
+        return  new ReactiveRedisTemplate<>(factory,context);
+    }
+
+}
