+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/models/enums/ProviderType.java
@@ -1,0 +1,6 @@
+package com.ds.news_fetcher_service.models.enums;
+
+public enum ProviderType {
+    RSS,
+    API
+}
