+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/models/provider/ProviderConfig.java
@@ -1,0 +1,17 @@
+package com.ds.news_fetcher_service.models.provider;
+
+import com.ds.news_fetcher_service.models.enums.ProviderType;
+import lombok.Data;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+public class ProviderConfig {
+    private String name;
+    private ProviderType type;
+    private String apiKeyNeeded;
+    private String apiKey;
+    private int rateLimitPerSecond;
+    private List<ProviderFeedConfig> feeds = new ArrayList<>();
+}
