+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/models/provider/ProviderFeedConfig.java
@@ -1,0 +1,9 @@
+package com.ds.news_fetcher_service.models.provider;
+
+import lombok.Data;
+
+@Data
+public class ProviderFeedConfig {
+    private String category;
+    private String url;
+}
