+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/provider/NewsProviderClient.java
@@ -1,0 +1,9 @@
+package com.ds.news_fetcher_service.provider;
+
+import com.ds.news_fetcher_service.models.payload.RawNewsPayload;
+import com.ds.news_fetcher_service.models.provider.ProviderConfig;
+import reactor.core.publisher.Flux;
+
+public interface NewsProviderClient {
+    Flux<RawNewsPayload> fetch(ProviderConfig provider);
+}
