+++ back/news-fetcher-service/src/main/java/com/ds/news_fetcher_service/util/HashUtil.java
@@ -1,0 +1,19 @@
+package com.ds.news_fetcher_service.util;
+
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.util.HexFormat;
+
+public class HashUtil {
+    private HashUtil() {}
+
+    public static String sha256(String value){
+        try{
+            MessageDigest md = MessageDigest.getInstance("SHA-256");
+            byte[] bytes = md.digest(value.getBytes(StandardCharsets.UTF_8));
+            return HexFormat.of().formatHex(bytes);
+        }catch(Exception e){
+            throw new RuntimeException(e);
+        }
+    }
+}
