+++ back/news-fetcher-service/src/test/java/com/ds/news_fetcher_service/NewsFetcherApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ds.news_fetcher_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NewsFetcherApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
