--- back/news-fetcher/src/main/java/com/ds/newsfetcher/NewsFetcherApplication.java
@@ -1,15 +1,0 @@
-package com.ds.newsfetcher;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-public class NewsFetcherApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NewsFetcherApplication.class, args);
-	}
-
-}
