--- back/news-fetcher/src/main/java/com/ds/newsfetcher/config/RedisConfig.java
@@ -1,22 +1,0 @@
-package com.ds.newsfetcher.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
-import org.springframework.data.redis.core.ReactiveRedisTemplate;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedisConfig {
-    @Bean
-    public ReactiveRedisTemplate<String, String> reactiveRedisTemplate(ReactiveRedisConnectionFactory factory){
-        RedisSerializationContext<String, String> context  = RedisSerializationContext
-                .<String, String>newSerializationContext(new StringRedisSerializer())
-                .value(new StringRedisSerializer())
-                .build();
-
-        return  new ReactiveRedisTemplate<>(factory,context);
-    }
-
-}
