--- back/news-fetcher/src/main/java/com/ds/newsfetcher/config/WebClientConfig.java
@@ -1,22 +1,0 @@
-package com.ds.newsfetcher.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    public WebClient.Builder webClientBuilder() {
-        return WebClient.builder();
-    }
-
-    @Bean
-    public WebClient webClient(WebClient.Builder builder) {
-        return builder
-                .codecs(clientCodecConfigurer -> clientCodecConfigurer.defaultCodecs().maxInMemorySize(5 * 1024 * 1024))
-                .build();
-    }
-
-}
