--- back/news-fetcher/src/main/java/com/ds/newsfetcher/models/enums/ProviderType.java
@@ -1,6 +1,0 @@
-package com.ds.newsfetcher.models.enums;
-
-public enum ProviderType {
-    RSS,
-    API
-}
