--- back/news-fetcher/src/main/java/com/ds/newsfetcher/models/provider/ProviderFeedConfig.java
@@ -1,9 +1,0 @@
-package com.ds.newsfetcher.models.provider;
-
-import lombok.Data;
-
-@Data
-public class ProviderFeedConfig {
-    private String category;
-    private String url;
-}
