--- back/news-fetcher/src/main/java/com/ds/newsfetcher/util/HashUtil.java
@@ -1,19 +1,0 @@
-package com.ds.newsfetcher.util;
-
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.util.HexFormat;
-
-public class HashUtil {
-    private HashUtil() {}
-
-    public static String sha256(String value){
-        try{
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            byte[] bytes = md.digest(value.getBytes(StandardCharsets.UTF_8));
-            return HexFormat.of().formatHex(bytes);
-        }catch(Exception e){
-            throw new RuntimeException(e);
-        }
-    }
-}
