--- back/news-fetcher/src/test/java/com/ds/newsfetcher/NewsFetcherApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ds.newsfetcher;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NewsFetcherApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
